+++ gfx/transform.go
@@ -71,7 +71,7 @@ func revertColor(rawColor uint8, index int, isPlus bool) color.Color {
 		}
 	} else {
 		plusColor := constants.NewRawCpcPlusColor(uint16(rawColor))
-		c := color.RGBA{A: 0xFF, R: uint8(plusColor.R), G: uint8(plusColor.G), B: uint8(plusColor.B)}
+		c := color.RGBA{A: 0xFF, R: plusColor.R, G: plusColor.G, B: plusColor.B}
 		newColor = constants.CpcPlusPalette.Convert(c)
 	}
 	return newColor
